+++ CHANGES.md
@@ -17,6 +17,10 @@ details.
 - Fix `deriving_inline` round-trip check so that it works with 5.01 <-> 5.02
   migrations (#519, @NathanReb)
 
+- Add ppxlib's AST pretty-printing utilities in `Ppxlib.Pp_ast` and
+  a `ppxlib-pp-ast` executable in a new separate `ppxlib-tools` package
+  (#517, @NathanReb)
+
 0.33.0 (2024-07-22)
 -------------------